LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scroll-markers/html-scroll-marker-target-before-after-expected.txt

First Second Third Fourth Fifth Sixth

FAIL 0th scroll marker test assert_equals: the :target-current scroll marker should be green expected "rgb(0, 128, 0)" but got "rgb(0, 0, 255)"
FAIL 1th scroll marker test assert_equals: scroll marker before the :target-current one should be red as :target-before expected "rgb(255, 0, 0)" but got "rgb(0, 0, 255)"
FAIL 2th scroll marker test assert_equals: scroll marker before the :target-current one should be red as :target-before expected "rgb(255, 0, 0)" but got "rgb(0, 0, 255)"
FAIL 3th scroll marker test assert_equals: scroll marker before the :target-current one should be red as :target-before expected "rgb(255, 0, 0)" but got "rgb(0, 0, 255)"
FAIL 4th scroll marker test assert_equals: scroll marker before the :target-current one should be red as :target-before expected "rgb(255, 0, 0)" but got "rgb(0, 0, 255)"
FAIL 5th scroll marker test assert_equals: scroll marker before the :target-current one should be red as :target-before expected "rgb(255, 0, 0)" but got "rgb(0, 0, 255)"
